[addons] move ./xbmc/addons/binary/interfaces one down and remove binary folder
