+++ lib/addons/library.xbmc.codec/libXBMC_codec.cpp
@@ -22,7 +22,7 @@
 #include <stdlib.h>
 #include <stdarg.h>
 #include <string>
-#include "addons/binary/interfaces/api1/Codec/AddonCallbacksCodec.h"
+#include "addons/interfaces/api1/Codec/AddonCallbacksCodec.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/libXBMC_codec.h"
 
 #ifdef _WIN32
