+++ project/cmake/treedata/common/subdirs.txt
@@ -1,14 +1,14 @@
 xbmc                                            xbmc
 xbmc/addons                                     addons
-xbmc/addons/binary/interfaces                   addonsBinaryInterfaces
-xbmc/addons/binary/interfaces/api1/Addon        api1AddonCallbacks_Addon
-xbmc/addons/binary/interfaces/api1/AudioDSP     api1AddonCallbacks_AudioDSP
-xbmc/addons/binary/interfaces/api1/AudioEngine  api1AddonCallbacks_AudioEngine
-xbmc/addons/binary/interfaces/api1/Codec        api1AddonCallbacks_Codec
-xbmc/addons/binary/interfaces/api1/GUI          api1AddonCallbacks_GUI
-xbmc/addons/binary/interfaces/api1/InputStream  api1AddonCallbacks_InputStream
-xbmc/addons/binary/interfaces/api1/Peripheral   api1AddonCallbacks_Peripheral
-xbmc/addons/binary/interfaces/api1/PVR          api1AddonCallbacks_PVR
+xbmc/addons/interfaces                          addonsBinaryInterfaces
+xbmc/addons/interfaces/api1/Addon               api1AddonCallbacks_Addon
+xbmc/addons/interfaces/api1/AudioDSP            api1AddonCallbacks_AudioDSP
+xbmc/addons/interfaces/api1/AudioEngine         api1AddonCallbacks_AudioEngine
+xbmc/addons/interfaces/api1/Codec               api1AddonCallbacks_Codec
+xbmc/addons/interfaces/api1/GUI                 api1AddonCallbacks_GUI
+xbmc/addons/interfaces/api1/InputStream         api1AddonCallbacks_InputStream
+xbmc/addons/interfaces/api1/Peripheral          api1AddonCallbacks_Peripheral
+xbmc/addons/interfaces/api1/PVR                 api1AddonCallbacks_PVR
 xbmc/commons                                    commons
 xbmc/cdrip                                      cdrip
 xbmc/dbwrappers                                 dbwrappers
