+++ xbmc/addons/interfaces/AddonInterfaces.cpp
@@ -23,15 +23,15 @@
 
 #include "addons/Addon.h"
 
-#include "addons/binary/interfaces/api1/Addon/AddonCallbacksAddon.h"
-#include "addons/binary/interfaces/api1/AudioDSP/AddonCallbacksAudioDSP.h"
-#include "addons/binary/interfaces/api1/AudioEngine/AddonCallbacksAudioEngine.h"
-#include "addons/binary/interfaces/api1/Codec/AddonCallbacksCodec.h"
-#include "addons/binary/interfaces/api1/GUI/AddonCallbacksGUI.h"
-#include "addons/binary/interfaces/api1/GUI/AddonGUIWindow.h"
-#include "addons/binary/interfaces/api1/InputStream/AddonCallbacksInputStream.h"
-#include "addons/binary/interfaces/api1/Peripheral/AddonCallbacksPeripheral.h"
-#include "addons/binary/interfaces/api1/PVR/AddonCallbacksPVR.h"
+#include "addons/interfaces/api1/Addon/AddonCallbacksAddon.h"
+#include "addons/interfaces/api1/AudioDSP/AddonCallbacksAudioDSP.h"
+#include "addons/interfaces/api1/AudioEngine/AddonCallbacksAudioEngine.h"
+#include "addons/interfaces/api1/Codec/AddonCallbacksCodec.h"
+#include "addons/interfaces/api1/GUI/AddonCallbacksGUI.h"
+#include "addons/interfaces/api1/GUI/AddonGUIWindow.h"
+#include "addons/interfaces/api1/InputStream/AddonCallbacksInputStream.h"
+#include "addons/interfaces/api1/Peripheral/AddonCallbacksPeripheral.h"
+#include "addons/interfaces/api1/PVR/AddonCallbacksPVR.h"
 #include "filesystem/SpecialProtocol.h"
 #include "messaging/ApplicationMessenger.h"
 #include "utils/log.h"
