+++ xbmc/addons/interfaces/AddonInterfaces.h
