+++ xbmc/addons/interfaces/CMakeLists.txt
