+++ xbmc/addons/interfaces/IAddonInterface.h
