+++ xbmc/addons/interfaces/api1/Addon/AddonCallbacksAddon.cpp
