+++ xbmc/addons/interfaces/api1/Addon/AddonCallbacksAddon.h
@@ -20,7 +20,7 @@
  *
  */
 
-#include "addons/binary/interfaces/AddonInterfaces.h"
+#include "addons/interfaces/AddonInterfaces.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/libXBMC_addon.h"
 
 struct VFSDirEntry;
