+++ xbmc/addons/interfaces/api1/Addon/CMakeLists.txt
