+++ xbmc/addons/interfaces/api1/Addon/Makefile
