+++ xbmc/addons/interfaces/api1/AudioDSP/AddonCallbacksAudioDSP.cpp
