+++ xbmc/addons/interfaces/api1/AudioDSP/AddonCallbacksAudioDSP.h
@@ -19,7 +19,7 @@
  *
  */
 
-#include "addons/binary/interfaces/AddonInterfaces.h"
+#include "addons/interfaces/AddonInterfaces.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/libKODI_adsp.h"
 
 namespace ActiveAE
