+++ xbmc/addons/interfaces/api1/AudioEngine/AddonCallbacksAudioEngine.cpp
