+++ xbmc/addons/interfaces/api1/GUI/AddonGUIRenderingControl.cpp
