+++ xbmc/addons/interfaces/api1/GUI/AddonGUIWindow.h
