+++ xbmc/addons/interfaces/api1/GUI/CMakeLists.txt
