+++ xbmc/addons/interfaces/api1/InputStream/AddonCallbacksInputStream.cpp
