+++ xbmc/addons/interfaces/api1/InputStream/AddonCallbacksInputStream.h
@@ -19,7 +19,7 @@
  *
  */
 
-#include "addons/binary/interfaces/AddonInterfaces.h"
+#include "addons/interfaces/AddonInterfaces.h"
 #include "cores/VideoPlayer/DVDDemuxers/DVDDemuxUtils.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/libKODI_inputstream.h"
