+++ xbmc/addons/interfaces/api1/InputStream/CMakeLists.txt
