+++ xbmc/addons/interfaces/api1/InputStream/Makefile
