+++ xbmc/addons/interfaces/api1/Peripheral/AddonCallbacksPeripheral.cpp
