+++ xbmc/addons/interfaces/api1/Peripheral/Makefile
